Fixing travis so it can build (#17)
* chore(CI): add module to the cache preparation (#9)

Add the module to the cache preparation. Although this is not necessary for most of the modules it is at least necessary for the eluna module. This PR is just about harmonizing the Travis configuration for all modules.

* fix: Include Chat.h so module would compile when NOPCH is used (#10)

Added include for Chat.h so module would compile when NOPCH was used (#10)
So when you take this module as a template, you don't miss any header :)

* docs: Add README.md in Spanish (#11)

* fix: Path to Config.h

Only need the file name.

* chore(CI): Update Travis configuration to use Ubuntu (#13)

* chore: Remove useless lines from README.md

* Adding support for db_assembler

* Update conf.sh.dist

* Fixing Travis

* Delete conf.sh.dist

Co-authored-by: Stoabrogga <38475780+Stoabrogga@users.noreply.github.com>
Co-authored-by: sluggor <50815505+sluggor@users.noreply.github.com>
Co-authored-by: Walter Pagani <paganiwalter@gmail.com>
Co-authored-by: Barbz <BarbzYHOOL@users.noreply.github.com>
Co-authored-by: Micrah <44911744+milestorme@users.noreply.github.com>